requirements-cpu.txt
altair==5.1.1
attrs==23.1.0
backports.zoneinfo;python_version<"3.9"
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.4.1
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.2
filelock==3.12.3
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.35
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lapx==0.5.4
lit==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
protobuf==4.24.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==13.0.0
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.10.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
streamlit==1.26.0
sympy==1.12
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
triton==2.0.0
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.173
urllib3==2.0.4
validators==0.22.0
watchdog==3.0.0
zipp==3.16.2
omegaconf==2.3.0
easyocr==1.7.0
imageio==2.31.5
